+++ vulture/core.py
@@ -136,7 +136,6 @@ class Vulture(ast.NodeVisitor):
         self.defined_vars = get_list('defined_vars')
         self.used_attrs = get_list('used_attrs')
         self.used_vars = get_list('used_vars')
-        self.names_imported_as_aliases = get_list('names_imported_as_aliases')
 
         self.filename = ''
         self.code = []
@@ -244,13 +243,13 @@ class Vulture(ast.NodeVisitor):
     def unused_classes(self):
         return _get_unused_items(
             self.defined_classes,
-            self.used_attrs + self.used_vars + self.names_imported_as_aliases)
+            self.used_attrs + self.used_vars)
 
     @property
     def unused_funcs(self):
         return _get_unused_items(
             self.defined_funcs,
-            self.used_attrs + self.used_vars + self.names_imported_as_aliases)
+            self.used_attrs + self.used_vars)
 
     @property
     def unused_imports(self):
@@ -266,7 +265,7 @@ class Vulture(ast.NodeVisitor):
     def unused_vars(self):
         return _get_unused_items(
             self.defined_vars,
-            self.used_attrs + self.used_vars + self.names_imported_as_aliases)
+            self.used_attrs + self.used_vars)
 
     @property
     def unused_attrs(self):
@@ -311,7 +310,7 @@ class Vulture(ast.NodeVisitor):
             self.defined_imports.append(
                 Item(alias or name, 'import', self.filename, node.lineno))
             if alias is not None:
-                self.names_imported_as_aliases.append(name_and_alias.name)
+                self.used_vars.append(name_and_alias.name)
 
     def _define_variable(self, name, lineno):
         if _ignore_variable(self.filename, name):