+++ NEWS.rst
@@ -7,11 +7,12 @@ News
 * Report attribute names that have the same names as variables as dead code.
 * Let Item class inherit from ``object`` (thanks @RJ722).
 * Handle names imported as aliases like all other used variable names.
-* Rename used_vars to used_names.
+* Rename Vulture.used_vars to Vulture.used_names.
 * Use function for determining which imports to ignore.
 * Only try to import each whitelist file once.
 * Store used names and used attributes as plain strings in sets instead of lists.
 * Fix estimating the size of code containing ellipses (...).
+* Refactor and simplify code.
 
 
 0.19 (2017-07-20)